+++ internal/api/signup.go
@@ -237,9 +237,6 @@ func (a *API) Signup(w http.ResponseWriter, r *http.Request) error {
 				}); terr != nil {
 					return terr
 				}
-				if terr = triggerEventHooks(ctx, tx, SignupEvent, user, config); terr != nil {
-					return terr
-				}
 				if terr = user.Confirm(tx); terr != nil {
 					return internalServerError("Database error updating user").WithInternalError(terr)
 				}
@@ -272,9 +269,6 @@ func (a *API) Signup(w http.ResponseWriter, r *http.Request) error {
 					"provider": params.Provider,
 					"channel":  params.Channel,
 				}); terr != nil {
-					return terr
-				}
-				if terr = triggerEventHooks(ctx, tx, SignupEvent, user, config); terr != nil {
 					return terr
 				}
 				if terr = user.ConfirmPhone(tx); terr != nil {
@@ -335,9 +329,6 @@ func (a *API) Signup(w http.ResponseWriter, r *http.Request) error {
 			if terr = models.NewAuditLogEntry(r, tx, user, models.LoginAction, "", map[string]interface{}{
 				"provider": params.Provider,
 			}); terr != nil {
-				return terr
-			}
-			if terr = triggerEventHooks(ctx, tx, LoginEvent, user, config); terr != nil {
 				return terr
 			}
 			token, terr = a.issueRefreshToken(ctx, tx, user, models.PasswordGrant, grantParams)
@@ -408,9 +399,6 @@ func (a *API) signupNewUser(ctx context.Context, conn *storage.Connection, user 
 		}
 		if terr = user.SetRole(tx, config.JWT.DefaultGroupName); terr != nil {
 			return internalServerError("Database error updating user").WithInternalError(terr)
-		}
-		if terr = triggerEventHooks(ctx, tx, ValidateEvent, user, config); terr != nil {
-			return terr
 		}
 		return nil
 	})