+++ internal/api/token.go
@@ -205,9 +205,6 @@ func (a *API) ResourceOwnerPasswordGrant(ctx context.Context, w http.ResponseWri
 		}); terr != nil {
 			return terr
 		}
-		if terr = triggerEventHooks(ctx, tx, LoginEvent, user, config); terr != nil {
-			return terr
-		}
 		token, terr = a.issueRefreshToken(ctx, tx, user, models.PasswordGrant, grantParams)
 		if terr != nil {
 			return terr
@@ -315,7 +312,7 @@ func (a *API) generateAccessToken(ctx context.Context, tx *storage.Connection, u
 		if terr != nil {
 			return "", 0, terr
 		}
-		aal, amr, terr = session.CalculateAALAndAMR(tx)
+		aal, amr, terr = session.CalculateAALAndAMR(user)
 		if terr != nil {
 			return "", 0, terr
 		}
@@ -452,12 +449,15 @@ func (a *API) updateMFASessionAndClaims(r *http.Request, tx *storage.Connection,
 		if terr != nil {
 			return terr
 		}
+		if err := tx.Load(user, "Identities"); err != nil {
+			return err
+		}
 		// Swap to ensure current token is the latest one
 		refreshToken, terr = models.GrantRefreshTokenSwap(r, tx, user, currentToken)
 		if terr != nil {
 			return terr
 		}
-		aal, _, terr := session.CalculateAALAndAMR(tx)
+		aal, _, terr := session.CalculateAALAndAMR(user)
 		if terr != nil {
 			return terr
 		}