+++ internal/api/verify.go
@@ -311,8 +311,6 @@ func (a *API) verifyPost(w http.ResponseWriter, r *http.Request, params *VerifyP
 }
 
 func (a *API) signupVerify(r *http.Request, ctx context.Context, conn *storage.Connection, user *models.User) (*models.User, error) {
-	config := a.config
-
 	if user.EncryptedPassword == "" && user.InvitedAt != nil {
 		// sign them up with temporary password, and require application
 		// to present the user with a password set form
@@ -338,10 +336,6 @@ func (a *API) signupVerify(r *http.Request, ctx context.Context, conn *storage.C
 			return terr
 		}
 
-		if terr = triggerEventHooks(ctx, tx, SignupEvent, user, config); terr != nil {
-			return terr
-		}
-
 		if terr = user.Confirm(tx); terr != nil {
 			return internalServerError("Error confirming user").WithInternalError(terr)
 		}
@@ -354,8 +348,6 @@ func (a *API) signupVerify(r *http.Request, ctx context.Context, conn *storage.C
 }
 
 func (a *API) recoverVerify(r *http.Request, ctx context.Context, conn *storage.Connection, user *models.User) (*models.User, error) {
-	config := a.config
-
 	err := conn.Transaction(func(tx *storage.Connection) error {
 		var terr error
 		if terr = user.Recover(tx); terr != nil {
@@ -366,17 +358,11 @@ func (a *API) recoverVerify(r *http.Request, ctx context.Context, conn *storage.
 				return terr
 			}
 
-			if terr = triggerEventHooks(ctx, tx, SignupEvent, user, config); terr != nil {
-				return terr
-			}
 			if terr = user.Confirm(tx); terr != nil {
 				return terr
 			}
 		} else {
 			if terr = models.NewAuditLogEntry(r, tx, user, models.LoginAction, "", nil); terr != nil {
-				return terr
-			}
-			if terr = triggerEventHooks(ctx, tx, LoginEvent, user, config); terr != nil {
 				return terr
 			}
 		}
@@ -390,12 +376,8 @@ func (a *API) recoverVerify(r *http.Request, ctx context.Context, conn *storage.
 }
 
 func (a *API) smsVerify(r *http.Request, ctx context.Context, conn *storage.Connection, user *models.User, params *VerifyParams) (*models.User, error) {
-	config := a.config
 
 	err := conn.Transaction(func(tx *storage.Connection) error {
-		if terr := triggerEventHooks(ctx, tx, SignupEvent, user, config); terr != nil {
-			return terr
-		}
 
 		if params.Type == smsVerification {
 			if terr := models.NewAuditLogEntry(r, tx, user, models.UserSignedUpAction, "", nil); terr != nil {
@@ -529,10 +511,6 @@ func (a *API) emailChangeVerify(r *http.Request, ctx context.Context, conn *stor
 			return terr
 		}
 
-		if terr := triggerEventHooks(ctx, tx, EmailChangeEvent, user, config); terr != nil {
-			return terr
-		}
-
 		if identity, terr := models.FindIdentityByIdAndProvider(tx, user.ID.String(), "email"); terr != nil {
 			if !models.IsNotFoundError(terr) {
 				return terr
@@ -540,14 +518,14 @@ func (a *API) emailChangeVerify(r *http.Request, ctx context.Context, conn *stor
 			// confirming the email change should create a new email identity if the user doesn't have one
 			if _, terr = a.createNewIdentity(tx, user, "email", structs.Map(provider.Claims{
 				Subject:       user.ID.String(),
-				Email:         params.Email,
+				Email:         user.EmailChange,
 				EmailVerified: true,
 			})); terr != nil {
 				return terr
 			}
 		} else {
 			if terr := identity.UpdateIdentityData(tx, map[string]interface{}{
-				"email":          params.Email,
+				"email":          user.EmailChange,
 				"email_verified": true,
 			}); terr != nil {
 				return terr